projects/travel-suite/apps/agents/requirements.txt
# Agno Multi-Agent Framework
agno>=0.1.0

# LLM Providers
openai>=1.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Vector DB Support (for RAG)
pgvector>=0.2.0

# Pydantic for type safety
pydantic>=2.0.0

# Error Tracking & Monitoring
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
respx>=0.20.0  # For mocking httpx requests
